refactor: isolate fuzz ecstore imports (#3762)
